Internal function to convert Proteomics Discoverer data.frame to SummarizedExperiment

Internal function to convert Proteomics Discoverer data.frame to SummarizedExperiment

convert_PD_df_to_SE(
  protein_df,
  ann_lib = c("org.Hs.eg.db"),
  curation_txt = NULL,
  ptm_colname = "Modifications",
  type = c("protein", "peptide"),
  remove_duplicate_peptides = TRUE,
  accession_from = NULL,
  accession_to = NULL,
  xref_df = NULL,
  verbose = FALSE,
  ...
)

Details

This function is intended to be called by import_proteomics_PD() after the Excel data is split into protein and peptide data.frame components.

See also