Convert data.frame to numeric color substitution

Convert data.frame to numeric color substitution

df_to_numcolors(
  df,
  colramp = c("Reds"),
  colramp_divergent = c("RdBu_r"),
  trimRamp = c(1, 2),
  ...
)

Arguments

df

data.frame where columns with numeric values will be colorized.

colramp, colramp_divergent

character name of color ramp for linear and divergent color gradients, respectively.

trimRamp

integer vector length=2, passed to jamba::getColorRamp(), to trim the edge colors from a linear color gradient, thus avoiding extreme colors.

...

additional arguments are ignored.

See also