Convert frequency matrix to normalizedMatrix format

Convert frequency matrix to normalizedMatrix format

frequency_matrix2nmat(
  mat,
  target_frequency = NULL,
  target_name = "target",
  signal_name = "Frequency",
  do_scale = TRUE,
  scale_from = 0,
  scale_to = 1,
  scale_frequency = NULL,
  apply_floor = TRUE,
  verbose = FALSE,
  ...
)

Arguments

mat

numeric matrix with frequency represented as columns, and whose numeric frequency is stored in colnames(mat) as character values.

target_frequency

numeric vector representing a range of frequencies to be considered the "target", and thus highlighted by EnrichedHeatmap::EnrichedHeatmap(). By default the target range is used to order rows from high to low signal, unless the order is specified otherwise. When target_frequency is NULL, there is no target indicated in the normalizedMatrix.

...

additional arguments are ignored.

Value

normalizedMatrix object, a subclass of matrix, as defined in EnrichedHeatmap.

Details

This function takes input data with frequency represented as columns, and observations as rows. It will optionally scale each row to have fixed minimum-to-maximum value range, given a range of frequencies to use for scaling.

See also