Get Salmon root directory

Get Salmon root directory

get_salmon_root(file)

Arguments

file

character path to a Salmon output file, or a vector of files.

Value

file path to the Salmon root directory, or NA if the Salmon root directory could not be found.

Details

This function uses the rprojroot package to find the root output directory of a Salmon quant analysis, looking for the file "cmd_info.json". If this file is not found, this function returns NULL.

See also

Examples

cmdinfopath <- system.file("data", "salmonOut", "cmd_info.json", package="platjam");
if (nchar(cmdinfopath) > 0) {
   get_salmon_root(cmdinfopath);
}
#> [1] "/private/var/folders/9z/y1_0kf711fx3lgc68w1k27xr0000gs/T/RtmpygQHlg/temp_libpath2987daa73a5/platjam/data/salmonOut"