Get UCSC track default values and track templates

Get UCSC track default values and track templates

get_track_defaults(env = new.env())

Arguments

env
environment in which to store the default values and track templates.

Value

environment which contains the default values, and template values.

Details

This function defines default values for overlay and composite track types. It also defines three templates each for overlay and composite:

overlay_header - equivalent to superTrack

overlay_parent - equivalent to one set of overlay tracks

overlay_track - each overlay track

composite_header - equivalent to one composite track

composite_parent - equivalent to a composite view within a composite track

composite_track - individual track within the composite track.

See also