Import LipoType lipidomics data from CSV file

Import LipoType lipidomics data from CSV file

import_lipotype_csv(
  csv,
  type = "raw",
  curation_txt = NULL,
  num_header_rows = 3,
  num_anno_columns = 7,
  verbose = FALSE,
  ...
)

Arguments

csv

character path to a comma-delimited file (csv) as provided in a LipoType summary report.

curation_txt

data.frame whose first column should match the column headers found in the LipoType measurement data. If curation_txt is not supplied, then values will be split into columns by _ underscore or " " whitespace characters.

...

additional arguments are ignored.

Value

SummarizedExperiment where:

rowData represents lipid measurement annotations

colData represents sample annotations, optionally including annotations via a data.frame supplied as curation_txt.

See also