Import Nanostring csv file

Import Nanostring csv file

import_nanostring_csv(
  csv,
  probe_colname = "Probe_ID",
  probe_anno_file = NULL,
  assay_name = NULL,
  hk_count = 10,
  curation_txt = NULL,
  verbose = FALSE,
  ...
)

Arguments

csv

character path to CSV file, suitable for import by data.table::fread(), so most delimiters will be recognized.

probe_colname

character string used to define the probe identifier in rowData(), derived from the colnames() of the import data. Sometimes the values represent "Probe_ID", and sometimes they may represent "Gene", "miRNA" or another measurement.

probe_anno_file

character path to one or more files with additional annotation for each measurement, with identifiers that match probe_colname above.

assay_name

character string of optional assay name used in the SummarizedExperiment object created. When NULL the default is "raw" unless the file contains substring "norm" in which case it uses "norm".

hk_count

integer number of housekeeper genes to use when "control_type" is not defined for the imported data. In this case, the last hk_count genes in the data are assumed to be housekeeper genes, by typical convention of NanoString codeset design.

curation_txt

either data.frame or character file path to tab- or comma-delimited file. The first column should match the column headers after importing data, colData(se). Subsequent columns contain associated sample annotations. For Nanostring data, the Nanostring sample annotations will already be associated with the colData(se), and colnames(curation_df) will overwrite any that already exist. Pro tip: The first column in curation_txt should contain '.' instead of punctuation/whitespace, to improve pattern matching filenames where the punctuation characters may have been modified during processing.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

SummarizedExperiment

See also