Import Nanostring RCC files

Import Nanostring RCC files

import_nanostring_rcc(
  rcc_files = NULL,
  rcc_path = ".",
  rcc_pattern = NULL,
  exclude = NULL,
  include = NULL,
  nprobes = -1,
  control_greps = c(POS = "^POS_", NEG = "^NEG_"),
  return_type = c("SummarizedExperiment", "NanoString"),
  debug = FALSE,
  verbose = FALSE,
  ...
)

Arguments

rcc_files

character vector of file paths to RCC files, or if NULL then the rcc_path and rcc_pattern arguments are used to find RCC files.

rcc_path

character vector of file directories to look for RCC files, the files are matched to pattern rcc_pattern. Note the rcc_path is only used when rcc_files is not supplied.

rcc_pattern

character string containing regular expression pattern, used to match filenames which should be considered RCC files.

exclude

optional character vector of filenames to exclude from import. The filenames may be either a full file path, or the basename without the file path.

include

optional character vector of filenames to include during data import. Note that when include is defined, only these files are included in the file import, all other files are ignored.

debug

logical indicating whether to send intermediate data before full processing, useful for debugging file format errors.

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

SummarizedExperiment or NanoString object, based upon argument return_type.

Details

This function reads one or more Nanostring RCC files, and produces a SummarizedExperiment object.

This function was based upon the function NanoStringNorm::read.markup.RCC(), which in our experience was problematic because it assumed all files contained identical sample annotation content.

For examples how to curate sample names into proper data.frame of experimental factors, see splicejam::curateVtoDF() or splicejam::curateDFtoDF().

See also