Import Nanostring RLF file for a codeset

Import Nanostring RLF file for a codeset

import_nanostring_rlf(
  rlf,
  plot_type = c("table", "list", "none"),
  color_sub = c(Endogenous = "darkorange", Positive = "greenyellow", Negative =
    "firebrick3", Housekeeping = "gold", Binding = "steelblue4", Purification =
    "steelblue2", Reserved = "mediumpurple1", ClassDate = "blue", ClassName_count =
    "navy"),
  desat = c(0.2, 0.4),
  ...
)

Arguments

rlf

character path to a Nanostring RLF file

plot_type

character string passed to design2colors() to define colors for the Class data. Use plot_type="none" to suppress plotting the colot table.

color_sub, desat

arguments passed to design2colors().

...

additional arguments are passed to design2colors().

Value

SummarizedExperiment object

Details

This function partially parses and imports data encoded in RLF file format by Nanostring for a particular Nanostring codeset. This file often includes information that links each probe to a "Class", which is then associated with other helpful information:

"ClassKey" - usually an integer unique identifier

"ClassName" - character string label for each ClassKey

"ClassActive" - integer value indicating the activity of probe for downstream stats.

"ClassDate" - the date associated with each class, usually identical for all classes in one codeset.

"ClassSource" - optional field describing the origin of each class

"ClassPreparer" - optional tool used to prepare classes in the RLF file

The most useful fields for downstream analysis are:

"ClassName" - which contains:

"Positive", "Negative" for the positive and negative control probes

"Endogenous" for the intended biological measurements, usually in the form of genes, transcripts, or miRNAs.

"Housekeeping" for the intended low-variance conrol measurements, usually in the form of genes, transcripts, miRNAs that are known or expected to have stable expression across the experiment conditions.

"Binding", "Purification", "Reserved" describe technical features beyond the scope of most analysis work.

"ClassActive" - which contains:

2 - a value that links "Endogenous" and "Housekeeping" above, though this value is not associated with another label described in the RLF file.

1 - a value that links "Binding", "Purification", "Reserved"

3 - a value that links "Positive" and "Negative" as control spike-in probes on the codeset

See also