Make multiple coverage heatmaps

Make multiple coverage heatmaps

nmatlist2heatmaps(nmatlist, k_clusters = 0, k_subset = NULL,
  k_colors = NULL, k_width = unit(3, "mm"), k_method = c("euclidean",
  "pearson", "correlation"), partition = NULL, rows = NULL,
  row_order = NULL, nmat_colors = NULL, main_heatmap = 1,
  anno_df = NULL, byCols = NULL, anno_row_marks = NULL,
  anno_row_labels = NULL, hm_nrow = 1, transform = jamba::log2signed,
  signal_ceiling = NULL, axis_name_gp = gpar(fontsize = 8),
  lens = -2, seed = 123, use_raster = TRUE, do_plot = TRUE,
  return_type = c("heatmaplist", "grid"), show_error = FALSE,
  verbose = FALSE, ...)

Arguments

nmatlist
list containing normalizedMatrix objects, usually the output from coverage_matrix2nmat().

k_clusters
integer number of k-means clusters to use to partition each heatmap. Use 0 or NULL for no clustering.

k_subset
integer vector of k-means clusters to retain. Often one cluster contains mostly empty values, and can be removed using this mechanism.

k_colors
vector of R colors, or NULL to use the output of colorjam::rainbowJam(k_clusters).

k_width
unit width of the k-means cluster color bar, used with k_clusters.

partition
vector used to split rows of each matrix in nmatlist, named by rownames. This value is ignored when k_clusters is supplied.

rows
vector of rownames or integer vector with index of rows to keep from each matrix in nmatlist.

row_order
integer vector used to order rows. When TRUE or NULL it uses the default for EnrichedHeatmap::EnrichedHeatmap() which is the EnrichedHeatmap::enriched_score() for the matrix main_heatmap. When FALSE the rows are ordered by the order they appear in rows, which is either the order they appear in nmatlist or the order after sorting anno_df. When TRUE the default

nmat_colors
named character vector of R colors, to colorize each heatmap. When NULL then colorjam::rainbowJam() is used to create colors for each heatmap panel.

main_heatmap
integer index referring to the entry in nmatlist to use for clustering and row ordering.

anno_df
data.frame or object that can be coerced, used to annotate rows of each matrix. It must have rownames(anno_df) that match rownames(nmatlist). When supplied, data can be sorted using byCols. Note that only the rownames(anno_df) present in both nmatlist and anno_df are used to display the heatmaps. These rows may also be subsetted using argument rows.

byCols
character vector of values in colnames(anno_df) used to sort the data.frame via jamba::mixedSortDF(). Any colname with prefix - will be reverse-sorted.

anno_row_marks
character vector of rownames which will be labeled beside the heatmaps, using the ComplexHeatmap::anno_mark() method. It currently requires anno_df be defined, since it uses the first column in anno_df as a one-column heatmap, to anchor the labels.

anno_row_labels
character vector of optional character labels to use instead of rownames. If NULL then anno_row_marks are used. Or anno_row_labels may contain a character vector of colnames(anno_df) which will create labels by concatenating each column value separated by space " ".

hm_nrow
integer number of rows used to display the heatmap panels.

transform
either character string referring to a numeric transformation, or a function that applies a numeric transformation. Valid character string values: "log2signed" applies jamba::log2signed() which applies log2(1+x) transform to the absolute value, then multiplies by the original sign(x); "sqrt" applies square root; "cubert" applies cube root x^(1/3); "qrt" applies fourth root x^(1/4). When there are negative numeric values, the transformation is applied to absolute value, then multiplied by the original sign. Therefore, the transformation is applied to adjust the magnitude of the values. These values are passed to get_numeric_transform() which may have more information.

signal_ceiling
numeric vector length length(nmatlist) which applies a maximum numeric value to the color ramp for each matrix in nmatlist. Every numeric value above the signal_ceiling will be assigned the maximum color. The values in signal_ceiling are recycled to length(nmatlist), so if one value is provided, it will be applied to every matrix.

axis_name_gp
x-axis label graphic parameters, as output from grid::gpar(). For example to define the x-axis font size, use the form grid::gpar(fontsize=8).

lens
numeric value used to scale each heatmap color ramp, using getColorRamp(). Values above zero apply the color gradient more rapidly starting from the lowest value, making the color appear more intense for lower numeric values. Values below zero apply the color gradient less rapidly, which makes lower numeric values appear less intense. This adjustment is intended to help apply suitable color contrast depending upon the range of numeric values. The lens values are applied to each matrix in nmatlist, and so it is recycled to length(nmatlist) as needed. Note that signal_ceiling is also intended to help apply the color gradient to a suitable numeric range, and the lens argument is applied relative to the numeric range being used.

seed
numeric value used with set.seed() to set the random seed. Set to NULL to avoid running set.seed().

use_raster
logical indicating whether to create heatmaps using raster resizing, almost always recommended TRUE.

do_plot
logical indicating whether to draw the heatmaps, where FALSE will return the data used to create heatmaps without actually drawing the heatmaps.

return_type
character string indicating the type of data to return: "heatmaplist" returns the list of heatmaps, which can separately be arranged together using ComplexHeatmap::draw() or grid::grid.draw().

show_error
logical indicating whether to add error bars to the profile plot at the top of each heatmap. These error bars are calculated by EnrichedHeatmap::anno_enriched() using matrixStats::colSds(x)/nrow(x).

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to EnrichedHeatmap::EnrichedHeatmap() to allow greater customization of details. Note that many ... arguments are also passed to ComplexHeatmap::Heatmap().

Details

This function takes a list of normalizedMatrix objects, usually the output of coverage_matrix2nmat(), and produces multiple heatmaps using EnrichedHeatmap.

This function is intended to be a convenient wrapper to help keep each data matrix in order, to apply consistent clustering and filtering across all data matrices, and to enable optional multi-row heatmap layout.

See also

Examples

## There is a small example file to use for testing
cov_file1 <- system.file("data", "tss_coverage.matrix", package="platjam");
cov_file2 <- system.file("data", "h3k4me1_coverage.matrix", package="platjam");
cov_files <- c(cov_file1, cov_file2);
names(cov_files) <- gsub("[.]matrix",
   "",
   basename(cov_files));
nmatlist <- lapply(cov_files, coverage_matrix2nmat);
nmatlist2heatmaps(nmatlist);

#> Error in get_numeric_transform(it, ...): object 'it' not found

# k-means clusters
nmatlist2heatmaps(nmatlist, k_clusters=4);

#> Error in get_numeric_transform(it, ...): object 'it' not found

# multiple rows
nmatlist2heatmaps(nmatlist, k_clusters=4, hm_nrow=2);

#> Error in get_numeric_transform(it, ...): object 'it' not found