Run contrasts on SummarizedExperiment data

Run contrasts on SummarizedExperiment data

run_contrasts_se(
  se,
  signalList = head(names(assays(se)), 1),
  cutoff_p = NULL,
  cutoff_adjp = 0.05,
  cutoff_fold = 1.5,
  cutoff_mgm = 0,
  cutoff_p_int = cutoff_p,
  cutoff_adjp_int = cutoff_adjp,
  cutoff_fold_int = cutoff_fold,
  confint = FALSE,
  probe_colname = NULL,
  gene_colname = "gene_name",
  idesign = NULL,
  icontrasts = NULL,
  enforce_design_subset = TRUE,
  igenes = NULL,
  isamples = NULL,
  use_voom = FALSE,
  weights = NULL,
  robust = FALSE,
  apply_probe_replicates = FALSE,
  floor_minimum = 0,
  floor_value = 0,
  fill_na = FALSE,
  na_weight = 0,
  return_lmfit = FALSE,
  return_df = TRUE,
  return_dfs = TRUE,
  return_hit_array = TRUE,
  collapse_by_gene = FALSE,
  stat_method = c("limma", "NanoStringDiff"),
  debug = 0,
  verbose = FALSE,
  ...
)

Arguments

se
SummarizedExperiment object with genes/proteins/probes in each row, and biological samples in each column.

cutoff_p, cutoff_adjp, cutoff_fold, cutoff_mgm
numeric vectors or NULL to define the thresholds to apply when defining statistical hits. When multiple values are supplied, each vector is recycled to the max length of these vectors. The hit criteria are applied for the first element in each vector, then the second element in each vector, and so on.

cutoff_p_int, cutoff_adjp_int, cutoff_fold_int
numeric vectors applied only to two-way or higher contrasts. For example, one may want to apply cutoff_adjp=0.01 for high stringency for simple pairwise contrasts, then cutoff_adjp_int=0.05 for less stringent criteria for two-way contrasts, due to the decrease in statistical power when testing twice as many groups.

confint
logical passed to limma::topTable() to indicate whether it returns upper and lower confidence intervals.

probe_colname, gene_colname
character values indicating the colname in rowData(se) that contains probe or gene information. Probe values are expected to represent unique probe assays, or unique transcript identified, or some equivalent definition. Gene values represent the gene or protein to which each individual entry is assigned, and is optional except when collapse_by_gene=TRUE.

idesign, icontrasts
numeric matrix to define the design and statistical contrasts. It is recommended that idesign has rownames that match colnames(se); and colnames(idesign) should match rownames(contrasts). The colnames(idesign) should not contain characters that interfere with contrast definition, for example no "-", "(", ")" characters. The colnames(icontrasts) can actually contain these characters, as this function makes effort to maintain the actual colnames(icontrasts) in the output for data integrity.

enforce_design_subset
logical indicating whether to use only the subset of samples represented in the full set of contrasts provided by icontrasts. For example, if testing only one two-group contrast from data that contains four groups, enforce_design_subset=TRUE will first reduce the input data to those two groups, then will run the limma workflow; enforce_design_subset=FALSE will keep all four groups, but only apply the contrast to the two groups. The latter process will use all four groups in the estimate of variance, which is applied to the two groups. The former process estimates variance only using the two groups provided.

igenes, isamples
character vectors with optional subset of genes which should match rownames(se), or samples which should match colnames(se). When all samples from a given sample group are removed, all associated contrasts are also removed.

use_voom
logical indicating whether to apply limma::voom() to define a weight matrix, which is used in the subsequent analysis steps.

weights
optional numeric matrix used in the limma analysis workflow, for example the output from limma::voom().

robust
logical passed to limma::eBayes() to indicate whether to apply robust estimate of variance, for example calling limma::squeezeVars().

apply_probe_replicates
logical indicating whether to combine probe replicates to one summary result per probe, only when there are multiple rows that represent the same probe (technical replicates.) When apply_probe_replicates=TRUE it applies the limma workflow with limma::duplicateCorrelation(), which is applied to every subset of probes with equivalent number of replicates. The set of probes which are not replicated is analyzed in its own batch, then all results are combined in the output. When apply_probe_replicates=TRUE the probe_colname argument should represent a column in rowData(se) that contains a repeated identifier that represents the technical replicates.

Value

list whose contents depend upon arguments: return_df=TRUE returns one large data.frame with all contrasts and stat hit flags applied; return_dfs=TRUE returns a list of data.frames named by contrast name, each include a summary stat table, with flags indicating stat hit and direction, the group mean, and max group mean values; return_array=TRUE will return an array with three dimensions: assay_name (the name of the assays analyzed from the input se object), contrast_name with each contrast, and hit_rule with the stat hit rules; return_lmfit=TRUE will return a list of lmFit as returned by limma::lmFit() and limma::fitContrasts().

Details

This function is not yet functional, it is being migrated from a private set of R functions. Until all required functions are migrated, this function is not usable.

This function is intended to combine several analysis steps to test statistical contrasts on data stored in a SummarizedExperiment object. Because SummarizedExperiment can store multiple assay data matrix entries, for example with data normalized by different methods, this function can also test each assay data and assembles results together for each comparison.

Optionally applies limma-voom workflow.

Applies statistical thresholds:

P-value, adjusted P-value

fold change

max group mean (highest mean of groups involved in each contrast)

Optionally applies a series of different stat hit thresholds

Optionally applies stat hits thresholds specific to two-way contrasts

Optionally combines results by gene, using collapse_by_gene()

Assembles a hit array, vectors named by gene or probe, whose values indicate the sign of the direction of statistical hits

Optionally applies recommended limma workflow for replicate microarray probe spots, for example limma::duplicateCorrelation() which returns one summary value per replicated probe

Optionally applies a signal floor, replacing values at or below the floor with the replacement (usually NA or the floor).

Optionally replaces NA values with an assigned value, usually 0 or the same floor value above.