Get appropriate numeric transformation function

Get appropriate numeric transformation function

get_numeric_transform(transform, ...)

Arguments

transform
character string or function. If length(transform) > 1 then get_numeric_transform() is called for each value in transform. Valid names: "log2signed" applies jamba::log2signed() which applies log2(1+x) transform to the absolute value, then multiplies by the original sign(x); "exp2signed" applies the inverse of "log2signed", which exponentiates numeric values which were previously transformed with log2(1+x); "sqrt" applies square root transform; "cubert" applies cube root x^(1/3); "qrt" applies fourth root x^(1/4); "frt" applies fourth root x^(1/5); "square" applies x^2 to absolute value, multiplied by the sign(x); "cube" applies x^3; When there are negative numeric values, the transformation is applied to absolute value, then multiplied by the original sign. Therefore, the transformation is applied to adjust the magnitude of the values.

...
additional arguments are ignored.

Value

function or NULL when no matching function is found, or list is returned when the input transform has multiple values.

Details

This function recognizes numeric transformation functions by name, or searches the attached R package environments for a matching function by name.

Examples

get_numeric_transform("log2signed")

#> Error in get_numeric_transform("log2signed"): object 'it' not found