Parse UCSC tracks that use the Gokey format

Parse UCSC tracks that use the Gokey format

parse_ucsc_gokey(
  track_lines,
  overlay_grep = c("[ -._](plus|minus|F|R|pos|neg)($|[ -._])"),
  priority = 5000,
  output_format = c("text", "list"),
  debug = c("none"),
  verbose = FALSE,
  ...
)

Arguments

track_lines

character vector containing lines read from a track file, or valid path or connection to a track file.

overlay_grep

character vector containing valid regular expression patterns used to recognize when a track should be considered an overlay coverage track. For example track name="trackA F" and track name="trackA R" would be recognized as forward and reverse strand for a track named "trackA". Overlay tracks are handled using the UCSC "multiWig" approach, and not the composite track approach. To disable overlay_grep, use overlay_grep="^$". To enable overlay_grep for all tracks, use overlay_grep="$".

priority

integer value indicating the priority to start when assigning priority to each track.

output_format

character string indicating the output format, where "text" will return one long character string, and "list" will return a list with one track per list element with class "glue","character".

debug

character indicating type of debug output: df returns the intermediate track_df data.frame; pri prints priority during track parsing; none does no debug.

...

additional arguments are ignored.

Value

by default a character string suitable to cat()directly into a text file, when output_format="text". When output_format="list" it returns a list of glue objects, which can be concatenated into one character string with Reduce("+", trackline_list).

Details

Given a text file, or lines from a text file, representing the Gokey format, this function will parse the track lines into groups, and return a text string usable in a UCSC genome browser track hub.

For example:

parse_ucsc_gokey("tracks.txt")

See also